requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
transformers>=4.30.0